refactor(capsule): extract CapsuleCache into its own file
