+++ scopes/component/isolator/index.ts
@@ -1,13 +1,8 @@
-export { CAPSULE_READY_FILE, CAPSULE_ORIGIN_FILE, CAPSULE_TRASH_DIR } from './isolator.main.runtime';
+export { CAPSULE_READY_FILE } from './isolator.main.runtime';
+export { CAPSULE_ORIGIN_FILE, CAPSULE_TRASH_DIR } from './capsule-cache';
 export { Network } from './network';
 export { FsContainer, Capsule, ContainerExec } from './capsule';
-export type {
-  IsolatorMain,
-  IsolateComponentsOptions,
-  CapsuleKind,
-  CapsuleOriginMarker,
-  PruneCapsulesOptions,
-  PruneCapsulesReport,
-} from './isolator.main.runtime';
+export type { IsolatorMain, IsolateComponentsOptions } from './isolator.main.runtime';
+export type { CapsuleKind, CapsuleOriginMarker, PruneCapsulesOptions, PruneCapsulesReport } from './capsule-cache';
 export { IsolatorAspect } from './isolator.aspect';
 export { default as CapsuleList } from './capsule-list';
